A way to acknowledge serialization as data preprocessing. https://online.sbis.ru/opendoc.html?guid=9f155dcd-4bed-46a7-bff6-9d638d2d284e
